refactor: simplified code and API due to moving of MessageDescriptor to generic.Client's KnownModel info retrieval data structure
